+++ CHANGELOG.md
@@ -10,12 +10,11 @@
 * not refresh graph while typing
 * upload texts mechanism (submit/accept)
 * bootstrap 5.3.8 widh CDN
-* weekly points
 
 
 ## 5.2.3 (2026-03-14)
 
-- 
+- weekly points - show count, sum(pt) in table format
 
 ## 5.2.2 (2026-03-14)